Menampilkan Detail Penghuni untuk menghapus

+++ app/src/main/java/com/proyektingkat2/villazone/db/PenghuniEntity.kt
@@ -13,6 +13,6 @@ data class PenghuniEntity(
     val namaPenghuni: String,
     val nomorHp: String,
     val nomorKamar: Int,
-    val biayaKamar: Int,
+    val biayaKamar: Double,
     val tanggalMasuk: String
 ) : Parcelable
